BUG/MEDIUM: autotable > Fix app error when no items
The App throws a bunch of errors if the array of tableItems returned by the API is null and can't be iterated. Also made fetchTableItems async to allow the app to wait for the data from the API before handling the response, thus removing additional potential errors. Also made minor syntax changes and changed calls and imports of Utils methods.