FEATURE: Create a lightweight Store plugin
Recently a lot of data was moved from the Vuex Store to the query params of the URL for persistency. The remaining store data doesn't require the full length of the Vuex plugin, so it was removed and a lightweight store plugin was created with some setters&getters for a few properties.